requirements.txt
asyncpg
python-dotenv
fastapi[standard]
apscheduler
asyncio
bcrypt